* commit '3b6169c332932f469fd84a671b351c31db8edebb':
  SW-18242 - add parameters to exceptions in media resource
  SW-18242 - Remove media file only if it was uploaded fixes https://github.com/shopware/shopware/pull/1047

+++ engine/Shopware/Models/Media/Media.php
@@ -2076,7 +2076,9 @@ class Media extends ModelEntity
             }
 
             $mediaService->write($this->path, file_get_contents($this->file->getRealPath()));
-            unlink($this->file->getPathname());
+            if (is_uploaded_file($this->file->getPathname())) {
+                unlink($this->file->getPathname());
+            }
         }
 
         return true;
